+++ android/src/main/kotlin/com/gdelataillade/alarm/alarm/AlarmService.kt
@@ -142,6 +142,7 @@ class AlarmService : Service() {
         audioService?.cleanUp()
         vibrationService?.stopVibrating()
         volumeService?.restorePreviousVolume(showSystemUI)
+        volumeService?.abandonAudioFocus()
 
         stopForeground(true)