+++ android/src/main/kotlin/com/gdelataillade/alarm/services/VolumeService.kt
@@ -6,10 +6,12 @@ import android.media.AudioAttributes
 import android.media.AudioFocusRequest
 import android.os.Build
 import kotlin.math.round
+import io.flutter.Log
 
 class VolumeService(private val context: Context) {
     private var previousVolume: Int? = null
     private val audioManager = context.getSystemService(Context.AUDIO_SERVICE) as AudioManager
+    private var focusRequest: AudioFocusRequest? = null
 
     fun setVolume(volume: Double, showSystemUI: Boolean) {
         val maxVolume = audioManager.getStreamMaxVolume(AudioManager.STREAM_MUSIC)
@@ -32,22 +34,40 @@ class VolumeService(private val context: Context) {
                 .setContentType(AudioAttributes.CONTENT_TYPE_SONIFICATION)
                 .build()
 
-            val focusRequest = AudioFocusRequest.Builder(AudioManager.AUDIOFOCUS_GAIN_TRANSIENT_MAY_DUCK)
+            focusRequest = AudioFocusRequest.Builder(AudioManager.AUDIOFOCUS_GAIN_TRANSIENT_MAY_DUCK)
                 .setAudioAttributes(audioAttributes)
                 .build()
 
-            audioManager.requestAudioFocus(focusRequest)
+            val result = audioManager.requestAudioFocus(focusRequest!!)
+            if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
+                Log.d("VolumeService", "Audio focus request granted")
+            } else {
+                Log.e("VolumeService", "Audio focus request failed")
+            }
         } else {
             @Suppress("DEPRECATION")
-            audioManager.requestAudioFocus(
+            val result = audioManager.requestAudioFocus(
                 null,
                 AudioManager.STREAM_ALARM,
                 AudioManager.AUDIOFOCUS_GAIN_TRANSIENT_MAY_DUCK
             )
+            if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
+                Log.d("VolumeService", "Audio focus request granted")
+            } else {
+                Log.e("VolumeService", "Audio focus request failed")
+            }
         }
     }
 
     fun abandonAudioFocus() {
-        audioManager.abandonAudioFocus(null)
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            focusRequest?.let {
+                audioManager.abandonAudioFocusRequest(it)
+            }
+        } else {
+            @Suppress("DEPRECATION")
+            audioManager.abandonAudioFocus(null)
+        }
+        Log.d("VolumeService", "Audio focus abandoned")
     }
 }